julia plots 라이브러리 사용 예제 추가

+++ JuliaPractice/helloworld/helloworld.jl
@@ -1,4 +1,4 @@
-function test( n::Int64 )
+function test( n )
     total = 0
 
     for i in 1:n
